Move merkle-tree cost calculation from hydra-node to dedicated bench in plutus-merkle-tree.
