+++ plutus-merkle-tree/bench/Validators.hs
@@ -6,11 +6,32 @@ module Validators where
 
 import PlutusTx.Prelude
 
-import Plutus.Extras (wrapValidator)
 import Plutus.MerkleTree (member)
 import qualified Plutus.MerkleTree as MT
-import Plutus.V1.Ledger.Api (Validator, mkValidatorScript)
+import Plutus.V1.Ledger.Api (ScriptContext, Validator, mkValidatorScript)
+import PlutusTx (UnsafeFromData (..))
 import qualified PlutusTx as Plutus
+
+-- | Signature of an untyped validator script.
+type ValidatorType = BuiltinData -> BuiltinData -> BuiltinData -> ()
+
+-- | Wrap a typed validator to get the basic `ValidatorType` signature which can
+-- be passed to `PlutusTx.compile`.
+-- REVIEW: There might be better ways to name this than "wrap"
+--
+-- TODO: Duplicate from Plutus.Extras, can't use because of circular deps.
+wrapValidator ::
+  (UnsafeFromData datum, UnsafeFromData redeemer) =>
+  (datum -> redeemer -> ScriptContext -> Bool) ->
+  ValidatorType
+-- We can use unsafeFromBuiltinData here as we would fail immediately anyway if parsing failed
+wrapValidator f d r p =
+  check $
+    f
+      (unsafeFromBuiltinData d)
+      (unsafeFromBuiltinData r)
+      (unsafeFromBuiltinData p)
+{-# INLINEABLE wrapValidator #-}
 
 -- | A validator for measuring cost of MT membership validation.
 merkleTreeMemberValidator :: Validator
